requirements.txt
face_recognition==1.3.0
opencv-python
numpy==1.26.4
python-dotenv
fastapi
uvicorn
sqlalchemy[asyncio]
psycopg2-binary
asyncpg
google-cloud-vision
